refactor(follow_setting): 标签功能ui迁移
