+++ simple_live_app/lib/modules/follow_user/follow_app_setting/follow_app_settings_controller.dart
@@ -1,7 +1,205 @@
-﻿import 'package:get/get.dart';
+﻿import 'package:flutter/material.dart';
+import 'package:flutter_smart_dialog/flutter_smart_dialog.dart';
+import 'package:get/get.dart';
+import 'package:simple_live_app/app/app_style.dart';
 import 'package:simple_live_app/app/controller/app_settings_controller.dart';
 import 'package:simple_live_app/app/controller/base_controller.dart';
+import 'package:simple_live_app/app/log.dart';
+import 'package:simple_live_app/app/utils.dart';
+import 'package:simple_live_app/models/db/follow_user_tag.dart';
+import 'package:simple_live_app/services/follow_service.dart';
 
-class FollowAppSettingsController extends BaseController{
+class FollowAppSettingsController extends BaseController {
   final appC = Get.find<AppSettingsController>();
+
+  // 用户自定义标签
+  RxList<FollowUserTag> userTagList = <FollowUserTag>[].obs;
+
+  @override
+  void onInit() {
+    updateTagList();
+    super.onInit();
+  }
+
+  // 标签管理
+  void updateTagList() {
+    userTagList.assignAll(FollowService.instance.followTagList);
+  }
+
+  Future removeTag(FollowUserTag tag) async {
+    await FollowService.instance.removeFollowUserTag(tag);
+    updateTagList();
+    Log.i('删除tag${tag.tag}');
+  }
+
+  void addTag(String tag) async {
+    await FollowService.instance.addFollowUserTag(tag);
+    updateTagList();
+  }
+
+  Future<void> updateTag(FollowUserTag followUserTag) async {
+    await FollowService.instance.updateFollowUserTag(followUserTag);
+  }
+
+  void updateTagName(FollowUserTag followUserTag, String newTagName) {
+    // 未操作
+    if (followUserTag.tag == newTagName) {
+      return;
+    }
+    // 避免重名
+    if (userTagList.any((item) => item.tag == newTagName)) {
+      SmartDialog.showToast("标签名重复，修改失败");
+      return;
+    }
+    FollowService.instance.updateTagName(followUserTag, newTagName);
+    SmartDialog.showToast("标签名修改成功");
+    updateTagList();
+  }
+
+  void updateTagOrder(int oldIndex, int newIndex) {
+    if (newIndex > oldIndex) newIndex -= 1; // 处理索引调整
+    final item = userTagList.removeAt(oldIndex);
+    userTagList.insert(newIndex, item);
+    FollowService.instance.updateFollowTagOrder(userTagList);
+  }
+
+  // 标签管理弹窗
+  void showTagsManager() {
+    Utils.showBottomSheet(
+      title: '标签管理',
+      child: Column(
+          mainAxisSize: MainAxisSize.min,
+          crossAxisAlignment: CrossAxisAlignment.stretch,
+          children: [
+            AppStyle.divider,
+            ListTile(
+              title: const Text("添加标签"),
+              leading: const Icon(Icons.add),
+              onTap: () {
+                editTagDialog("添加标签");
+              },
+            ),
+            AppStyle.divider,
+            // 列表内容
+            Expanded(
+              child: Obx(
+                () => ReorderableListView.builder(
+                  buildDefaultDragHandles: false,
+                  itemCount: userTagList.length,
+                  itemBuilder: (context, index) {
+                    // 偏移
+                    FollowUserTag item = userTagList[index];
+                    return ListTile(
+                      key: ValueKey(item.id),
+                      title: GestureDetector(
+                        child: Text(item.tag),
+                        onLongPress: () {
+                          {
+                            editTagDialog("修改标签", followUserTag: item);
+                          }
+                        },
+                      ),
+                      leading: IconButton(
+                        icon: const Icon(Icons.delete),
+                        onPressed: () {
+                          removeTag(item);
+                        },
+                      ),
+                      trailing: ReorderableDelayedDragStartListener(
+                        index: index,
+                        child: const Padding(
+                          padding: EdgeInsets.symmetric(horizontal: 12.0),
+                          child: Icon(Icons.drag_handle),
+                        ),
+                      ),
+                    );
+                  },
+                  onReorder: (int oldIndex, int newIndex) {
+                    updateTagOrder(oldIndex, newIndex);
+                  },
+                ),
+              ),
+            ),
+          ]),
+    );
+  }
+
+  void editTagDialog(String title, {FollowUserTag? followUserTag}) {
+    final TextEditingController tagEditController =
+        TextEditingController(text: followUserTag?.tag);
+    bool upMode = title == "添加标签" ? true : false;
+    Get.dialog(
+      AlertDialog(
+        contentPadding: const EdgeInsets.all(16.0),
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(12.0),
+        ),
+        content: SingleChildScrollView(
+          padding: EdgeInsets.only(
+            bottom: MediaQuery.of(Get.context!).viewInsets.bottom,
+          ),
+          child: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              Text(
+                title,
+                style: const TextStyle(
+                  fontSize: 18,
+                ),
+              ),
+              TextField(
+                controller: tagEditController,
+                minLines: 1,
+                maxLines: 1,
+                decoration: InputDecoration(
+                  border: const OutlineInputBorder(),
+                  contentPadding: AppStyle.edgeInsetsA12,
+                  enabledBorder: OutlineInputBorder(
+                    borderSide: BorderSide(
+                      color: Colors.grey.withValues(
+                        alpha: .2,
+                      ),
+                    ),
+                  ),
+                ),
+                onSubmitted: (tag) {
+                  upMode
+                      ? addTag(tagEditController.text)
+                      : updateTagName(followUserTag!, tagEditController.text);
+                  Get.back();
+                },
+              ),
+              Container(
+                margin: AppStyle.edgeInsetsB4,
+                width: double.infinity,
+                child: Row(
+                  mainAxisAlignment: MainAxisAlignment.end,
+                  children: [
+                    TextButton(
+                      onPressed: () {
+                        Get.back();
+                      },
+                      child: const Text('否'),
+                    ),
+                    TextButton(
+                      onPressed: () {
+                        upMode
+                            ? addTag(tagEditController.text)
+                            : updateTagName(
+                                followUserTag!,
+                                tagEditController.text,
+                              );
+                        Get.back();
+                      },
+                      child: const Text('是'),
+                    ),
+                  ],
+                ),
+              )
+            ],
+          ),
+        ),
+      ),
+    );
+  }
 }
